% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{CI.classNRI}
\alias{CI.classNRI}
\title{Statistical metrics and confidence intervals for classes}
\usage{
CI.classNRI(
  c1,
  c2,
  y,
  s1 = NULL,
  s2 = NULL,
  conf.level = 0.95,
  n.boot = 1000,
  dp = 3
)
}
\arguments{
\item{c1}{Risk classes of the baseline model (ordinal)}

\item{c2}{Risk classes of new model}

\item{y}{Binary of outcome of interest. Must be 0 or 1.}

\item{s1}{The savings or benefit when am event is reclassified to a higher group by the new model (positive numeric)}

\item{s2}{The benefit when a non-event is reclassified to a lower group (positive numeric)}

\item{conf.level}{The confidence interval expressed as a fraction of 1 (ie 0.95 is the 95\% confidence interval )}

\item{n.boot}{The number of "bootstraps" to use. Performance slows down with more bootstraps. For trialling result, use a low number (eg 2), for accuracy use a large number (eg 2000)}

\item{dp}{The number of decimal places to display}
}
\value{
A list with four items: 
\itemize{
\item{\strong{1. meta_data}} {Some overall meta data - Confidence Interval, number of bootstraps, s1, s2}
\item{\strong{2. Metrics}} {Point estimates of the statistical metrics (see list below)}
\item{\strong{3. Each_bootstrap_metrics}} {Point estimates of the statistical metrics for each bootstrapped sample (see list below)}
\item{\strong{4. Summary Metrics}} {Point estimates with confidence intervals of the statistical metrics. See following list:)}
\itemize{
 \item{\strong{Total (n)}} {Total number of subjects}
\item{\strong{Events (n)}} {Number of subjects with the event (outcome) of interest}
\item{\strong{Non-events (n)}} {Number of subjects without the event (outcome) of interest}
\item{\strong{Prevalence}} {The prevalence of the event}
\item{\strong{NRI events}} {The NRI with confidence interval for those with the event.}
\item{\strong{NRI non-events}} {The NRI with confidence interval for those without the event.}
\item{\strong{wNRI-events}} {The weighted NRI for those with the event}
\item{\strong{wNRI-nonevents}} {The weighted NRI for those without the event }
\item{\strong{Confusion matrix events}} {A confusion matric (table) showing the relationship of the numbers classified into each class with the baseline model compared to those in each class with the new model}
\item{\strong{Confusion matrix non-events}} {A confusion matric (table) showing the relationship of the numbers classified into each class with the baseline model compared to those in each class with the new model}
}
}

A matrix of metrics
}
\description{
The function CI.classNRI calculates the NRI statistics for reclassification of data already in classes with confidence intervals.  Uses statistics.classNRI.
}
