% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{clean_insurance_data}
\alias{clean_insurance_data}
\title{INSURANCE dataset}
\usage{
clean_insurance_data(insurance, col1_of_interest, col2_of_interest)
}
\arguments{
\item{insurance}{A dataframe consisting the information about the insurance.We only select those person whose charges are  more than 0.}

\item{col1_of_interest}{The user can select a column of their choice to check different things about the insured person}

\item{col2_of_interest}{The users wish is granted and a second column can be chosen to be investigated.}
}
\value{
A simplified dataframe
}
\description{
INSURANCE dataset
}
\details{
This function returns a very simplified version of the original dataset, such that the person who is interested in studying the insurance dataset is not
overwhelmed
}
\examples{
clean_insurance_data(insurance, age, charges)
}
