% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap_data.R
\docType{data}
\name{data_class}
\alias{data_class}
\title{Simple data set with classifications}
\format{
data frame with 3 columns
\describe{
  \item{ref_class}{The class (in this case as an interger) of the baseline model}
  \item{new_class}{The class (in this case as an interger) of the new model}
  \item{outcome}{The outcome of interest (0 or 1)}
}
}
\usage{
data_class
}
\description{
Example data for use with CI.classNRI
}
\keyword{datasets}
