\name{geneDescriptionInfo}
\alias{geneDescriptionInfo}
\docType{data}



\title{
  Description information of 27,416 protein-coding genes in Arabidopsis thaliana.
}


\description{
  Descriptive information including locus, description and Wiki description in Arabidopsis thaliana.
}


\usage{
  data(geneDescriptionInfo)
}



\author{
  Jingjing Zhai, Chuang Ma
}


\keyword{ data }

