% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{ggcontribute}
\alias{ggcontribute}
\title{The Contribution plot}
\usage{
ggcontribute(x1, x2 = NULL, option_flag = c("chi2", "percent"))
}
\arguments{
\item{x1}{Either a logistic regression fitted using glm (base package) or lrm (rms package)  of the baseline model.}

\item{x2}{Either a logistic regression fitted using glm (base package) or lrm (rms package)  of the new (alternative) model.}

\item{option_flag}{A flag to choose if the relative percentage of the Chi2-degrees of freedom are plotted.}
}
\description{
ggcontribute plots the contribution of each variable to the model
}
