% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Medical Insurance Dataset}
\format{

}
\source{
<https://www.kaggle.com/datasets/mirichoi0218/insurance?resource=download>
}
\usage{
insurance
}
\description{
A tidy dataset of Medical cost personal.
}
\keyword{datasets}
