% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{wap}
\alias{wap}
\alias{nap}
\alias{lap}
\alias{rap}
\title{Map over columns of a data frame simultaneously}
\usage{
wap(.tbl, .f)

nap(.tbl, .f)

lap(.tbl, .f)

rap(.tbl, ...)
}
\arguments{
\item{.tbl}{A data frame}

\item{.f}{a single formula}

\item{...}{named formulas

The \emph{lhs} of each formula indicates the type, in the \code{\link[vctrs:vec_c]{vctrs::vec_c()}} sense.
\itemize{
\item empty or \code{list()}: no check is performed on the results of
the rhs expressionand a list is returned.
\item \code{data.frame()}:  to indicate that the rhs should evaluate
to a data frame of 1 row. The data frames don't need to be of a specific types
and are are combined with \code{\link[vctrs:vec_rbind]{vctrs::vec_rbind()}}.
\item A data frame of a specific type, e.g. \code{data.frame(x = integer(), y = double())}
The rhs should evaluate to a data frame of that type with 1 row.
\item Any other ptype that makes sense for \code{\link[vctrs:vec_c]{vctrs::vec_c()}}. Each result must
validate \code{vctrs::vec_size(.) == 1L} and are combined with
\code{vctrs::vec_c(!!!, .ptype = .ptype)}
}

The rhs of each formula uses columns of \code{.tbl}, and each stands for a single
observation.}
}
\value{
\itemize{
\item \code{wap()} returns a vector of the type specified by the lhs of the formula.
The vector validates \code{vec_size() == nrow(.tbl)}. This is similar
to \code{\link[purrr:pmap]{purrr::pmap()}}
\item \code{nap()} returns \emph{n}othing, and can be used for side effects, similar to \code{\link[purrr:pwalk]{purrr::pwalk()}}
\item \code{rap()} adds a column to \code{.tbl} per formula in \code{...}
}
}
\description{
Map over columns of a data frame simultaneously
}
\examples{

library(purrr)
library(dplyr)
library(tibble)

tbl <- tibble(cyl = c(4, 6, 8), mpg = c(30, 25, 20))

# returns a list of 3 elements
tbl \%>\%
  wap(~ filter(mtcars, cyl == !!cyl, mpg < !!mpg))

# same
tbl \%>\%
  wap(list() ~ filter(mtcars, cyl == !!cyl, mpg < !!mpg))

# can specify the output type with the formula lhs
tbl \%>\%
  wap(integer() ~ nrow(filter(mtcars, cyl == !!cyl, mpg < !!mpg)))

# to make data frames
starwars \%>\%
  wap( data.frame() ~
    data.frame(species = length(species), films = length(films))
  )

# rap adds columns
tbl \%>\%
  rap(
     x =           ~ filter(mtcars, cyl == !!cyl, mpg < !!mpg),
     n = integer() ~ nrow(x)
  )

# rap is especially useful for iterating
# over multiple models
starwars \%>\%
  nest_by(gender) \%>\%
  rap(
    model = ~lm(height ~ mass + birth_year, data = data),
    perf  = double() ~ summary(model)$adj.r.squared
  )

}
