% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rap.R
\name{raplot}
\alias{raplot}
\title{Risk Assessment Plot.}
\usage{
raplot(x1, x2, y)
}
\arguments{
\item{x1}{Calculated probabilities (eg through a logistic regression model) of the reference (null) model.  Must be between 0 & 1}

\item{x2}{Calculated probabilities (eg through a logistic regression model) of the new (alternative) model.   Must be between 0 & 1}

\item{y}{Binary of outcome of interest. Must be 0 or 1.}
}
\value{
A list is returned with the following items:
\item{Null.p.sens}{The horizontal axis coordinates (calculated probablities) for the sensitivity of the reference (null) model}
 \item{Null.sens}{The vertical axis coordinates (sensitivities) of the reference (null) model}
  \item{Null.p.1spec}{The horizontal axis coordinates (calculated probablities) for 1-specificity of the reference (null) model}
   \item{Null.1spec}{The vertical axis coordinates (1-specificities) of the reference (null) model}
\item{Alt.p.sens}{The horizontal axis coordinates (calculated probablities) for the sensitivity of the new (alternative) model}
 \item{Alt.sens}{The vertical axis coordinates (sensitivities) of the new (alternative) model}
  \item{Alt.p.1spec}{The horizontal axis coordinates (calculated probablities) for 1-specificity of the new (alternative) model}
   \item{Alt.1spec}{The vertical axis coordinates (1-specificities) of the new (alternative) model}
}
\description{
The function raplots plots the Sensitivity and 1-Specificity curves against the calculated risk for the null (reference) and alternative (new) models, thus graphically displaying the IDIs for those with and without the events.  These plots can aid interpretation of the NRI and IDI metrics.
}
\section{Further reading }{
 The Risk Assessment Plot in this form was described by Pickering, J. W., & Endre, Z. H. (2012). New Metrics for Assessing Diagnostic Potential of Candidate Biomarkers. Clinical Journal of the American Society of Nephrology, 7, 1355–1364. doi:10.2215/CJN.09590911
}
\examples{
\dontrun{
data(data_risk)
outcome<-data_risk$outcome
null_mod<-data_risk$ref
alt_mod<-data_risk$new
output<-raplot(null_mod,alt_mod,outcome)
}
}

