% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{slam}
\alias{slam}
\title{pmap adapter}
\usage{
slam(.tbl, formula)
}
\arguments{
\item{.tbl}{a tibble}

\item{formula}{a formula that uses columns from the tibble}
}
\description{
\code{slam()} is typically used in \code{\link[purrr:pmap]{purrr::pmap()}} calls to transform a
formula that uses the raw names into a function. \code{slam()} is not quite
\code{\link[=rap]{rap()}}.
}
\examples{
library(purrr)
library(gapminder)
library(dplyr)

oceania <- gapminder::gapminder \%>\%
  filter(continent == "Oceania") \%>\%
  mutate(yr1952 = year - 1952) \%>\%
  select(-continent) \%>\%
  group_nest(country)

# the idea of slam is to promote a formula expressed with the
# column names into a pmap() ready function
formula <- ~broom::tidy(stats::lm(lifeExp ~ yr1952, data))
oceania \%>\%
  pmap(slam(oceania, formula))

# this is similar to e.g.
oceania \%>\%
  wap(~broom::tidy(stats::lm(lifeExp ~ yr1952, data)))

}
