% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{summary.rap}
\alias{summary.rap}
\title{List risk assessment metrics}
\usage{
\method{summary}{rap}(l)
}
\arguments{
\item{l}{List returned from CI.raplot}
}
\value{
A tibble
}
\description{
Display the summary metrics
}
