% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rap.R
\name{CI.classNRI}
\alias{CI.classNRI}
\title{Reclassification metrics  and confidence intervals with classes (ordinals) as inputs}
\usage{
CI.classNRI(c1, c2, y, s1, s2, cis = c("asymptotic", "boot"),
  conf.level = 0.95, n.boot = 2000, dp = 4)
}
\arguments{
\item{c1}{Risk class of Reference model (ordinal)}

\item{c2}{Risk class of Reference model (new)}

\item{y}{Binary of outcome of interest. Must be 0 or 1.}

\item{s1}{The savings or benefit when am event is reclassified to a higher group by the new model}

\item{s2}{The benefit when a non-event is reclassified to a lower group}

\item{cis}{The confidence interval method to use "asymptotic" or "boot".  Normally "boot" is used}

\item{conf.level}{The confidence interval expressed as a fraction of 1 (ie 0.95 is the 95\% confidence interval )}

\item{n.boot}{The number of "bootstraps" to use. Performance slows down with more bootstraps. For trialling result, use a low number (eg 2), for accuracy use a large number (eg 2000)}
}
\value{
A matrix is returned with the following items:
\item{Total (n)}{Total number of subjects}
 \item{Events (n)}{Number of subjects with the event (outcome) of intrest}
  \item{Non-events (n)}{Number of subjects without the event (outcome) of intrest}
   \item{NRI events}{The NRI with confidence interval for those with the event.}
\item{NRI non-events}{The NRI with confidence interval for those without the event.}
\item{NRI}{The sum of NRI events and NRI non-events. This is between -200 and +200}
\item{wNRI}{The weighted NRI}
}
\description{
The function CI.classNRI calculates the NRI statistics for reclassification of data already in classes with confidence intervals.  Uses statistics.classNRI.
}
\examples{
\dontrun{
data(data_class)
y<-data_class$outcome
c1<-data_class$ref_class
c2<-data_class$new_class
T=0.1 # e.g.
s1=1/T
s2=1/(1-T)
output<-CI.classNRI(c1, c2, y,s1,s2,cis = c("boot"), conf.level = 0.95, n.boot = 4, dp = 4)
}
}

