% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rap.R
\name{CI.raplot}
\alias{CI.raplot}
\title{Reclassification metrics and confidence intervals with calculated risk as inputs .}
\usage{
CI.raplot(x1, x2, y, s1, s2, t, cis = c("asymptotic", "boot"),
  conf.level = 0.95, n.boot = 2000, dp = 4)
}
\arguments{
\item{x1}{Calculated probabilities (eg through a logistic regression model) of the reference (null) model.  Must be between 0 & 1}

\item{x2}{Calculated probabilities (eg through a logistic regression model) of the new (alternative) model.   Must be between 0 & 1}

\item{y}{Binary of outcome of interest. Must be 0 or 1.}

\item{s1}{The savings or benefit when am event is reclassified to a higher group by the new model}

\item{s2}{The benefit when a non-event is reclassified to a lower group}

\item{t}{The risk threshold(s) for groups. eg t<-c(0,0.1,1) is a two group model with a threshold of 0.1 & t<-c(0,0.1,0.3,1)  is a three group model with thresholds at 0.1 and 0.3.}

\item{cis}{The confidence interval method to use "asymptotic" or "boot".  Normally "boot" is used}

\item{conf.level}{The confidence interval expressed as a fraction of 1 (ie 0.95 is the 95\% confidence interval )}

\item{n.boot}{The number of "bootstraps" to use. Performance slows down with more bootstraps. For trialling result, use a low number (eg 2), for accuracy use a large number (eg 2000)}
}
\value{
A matrix is returned with the following items:
\item{Total (n)}{Total number of subjects}
 \item{Events (n)}{Number of subjects with the event (outcome) of intrest}
  \item{Non-events (n)}{Number of subjects without the event (outcome) of intrest}
   \item{cfNRI events}{The category free (continuous) NRI (Net Reclassification Improvement) with confidence interval for those with the event. While expressed as a percentage this is between -200 and +200}
\item{cfNRI non-events}{The category free (continuous) NRI with confidence interval for those without the event. While expressed as a percentage this is between -200 and +200}
\item{cfNRI}{The sum of cfNRI events and cfNRI non-events}
\item{NRI events}{The NRI with confidence interval for those with the event.}
\item{NRI non-events}{The NRI with confidence interval for those without the event.}
\item{NRI}{The sum of NRI events and NRI non-events. This is between -200 and +200}
\item{wNRI}{The weighted NRI}
\item{IDI events}{The IDI (Integrated Discrimination Improvement) with confidence interval for those with the event. Expressed as a fraction}
\item{IDI non-events}{The IDI with confidence interval for those without the event. Expressed as a fraction}
\item{IDI}{The sum of IDI events and IDI non-events}
\item{relIDI}{The relative IDI}
\item{IS(null model)}{The Integrated Sensitivity (area under the sensitivity-calculated risk curve) for the reference (null) model}
\item{IS(alt model)}{The Integrated Sensitivity for the reference (alt) model. Note, the IDI events should be the difference between IS(alt model) and IS(null model)}
\item{IP(null model)}{The Integrated 1-Specificity (area under the 1-specificity-calculated risk curve) for the reference (null) model}
\item{IP(alt model)}{The Integrated S1-Specificity for the reference (alt) model. Note, the IDI non-events should be the difference between IP(alt model) and IP(null model)}
\item{AUC(null model)}{The Area Under the Receiver Operator Characteristic Curve for the reference (null) model}
\item{AUC(alt model)}{The Area Under the Receiver Operator Characteristic Curve for the reference (alt) model}
\item{difference (p)}{P value for the difference in AUCs (DeLong method)}
\item{incidence}{The incidence of the event}
}
\description{
The CI.raplot function produces summary metrics for risk assessment. Outputs the NRI, IDI, weighted NRI and category Free NRI all for those with events and those without events.  Also the AUCs of the two models and the comparison (DeLong) between AUCs. Output includes confidence intervals. Uses statistics.raplot.  Displayed graphically by raplot.
}
\section{Further reading }{

A useful reference for the weighted NRI is: Van Calster, B., Vickers, A. J., Pencina, M. J., Baker, S. G., Timmerman, D., & Steyerberg, E. W. (2013). Evaluation of markers and risk prediction models: overview of relationships between NRI and decision-analytic measures. Medical Decision Making : an International Journal of the Society for Medical Decision Making, 33(4), 490–501. doi:10.1177/0272989X12470757
Reference for the NRI and IDI: Pencina, M. J., D'Agostino, R. B., & Vasan, R. S. (2008). Evaluating the added predictive ability of a new marker: From area under the ROC curve to reclassification and beyond. Statistics in Medicine, 27(2), 157–172. doi:10.1002/sim.2929
Reference for the continuous (category free) NRI: Pencina, M. J., D'Agostino, R. B., & Steyerberg, E. W. (2011). Extensions of net reclassification improvement calculations to measure usefulness of new biomarkers. Statistics in Medicine, 30(1), 11–21. doi:10.1002/sim.4085
Reference for the method of comparing AUCs: DeLong, E., DeLong, D., & Clarke-Pearson, D. (1988). Comparing the areas under 2 or more correlated receiver operating characteristic curves - a nonparametric approach. Biometrics, 44(3), 837–845.
}
\examples{
\dontrun{
data(data_risk)
y<-data_risk$outcome
x1<-data_risk$ref
x2<-data_risk$new
t<-c(0,0.19,1)
T=0.1 #e.g.
s1=1/T
s2=1/(1-T)
output<-CI.raplot(x1, x2, y,s1,s2,t,cis = c("boot"), conf.level = 0.95, n.boot = 4, dp = 4)
}
}

