% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{ggcalibrate}
\alias{ggcalibrate}
\title{The Calibration plot}
\usage{
ggcalibrate(x1, x2 = NULL, y = NULL, n_knots = 5, ci_level = 0.95)
}
\arguments{
\item{x1}{Either a logistic regression fitted using glm (base package) or lrm (rms package) or calculated probabilities (eg through a logistic regression model) of the baseline model.  Must be between 0 & 1}

\item{x2}{Either a logistic regression fitted using glm (base package) or lrm (rms package) or calculated probabilities (eg through a logistic regression model) of the new (alternative) model.   Must be between 0 & 1}

\item{y}{Binary of outcome of interest. Must be 0 or 1 (if fitted models are provided this is extracted from the fit which for an rms fit must have x = TRUE, y = TRUE).}

\item{n_knots}{The curves are made by fitting a restricted cubic spline (rms package). The default 5-knots is usually enough.}

\item{ci_level}{Confidence interval of the curve (default = 0.95).}

\item{models}{One of three strings:
\itemize{
\item{\strong{both}} {plot(s) will contain both the Baseline and New models;} 
\item{\strong{x1}} {plot(s) will contain only the Baseline model;} 
\item{\strong{x2}} {plot(s) will contain only the New model.}
}}
}
\value{
a ggplot
}
\description{
ggcalibrate plots the stats::predicted events against the actual event rate
}
\examples{
\dontrun{
data(data_risk)
y<-data_risk$outcome 
x1<-data_risk$baseline
x2<-data_risk$new
#e.g.
output <- ggcalibrate(x1, x2, y, models = "both") 
}
}
