% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rap.R
\name{statistics.classNRI}
\alias{statistics.classNRI}
\title{Reclassification metrics with classes (ordinals) as inputs.}
\usage{
statistics.classNRI(c1, c2, y, s1, s2)
}
\arguments{
\item{c1}{Risk class of Reference model (ordinal)}

\item{c2}{Risk class of Reference model (new)}

\item{y}{Binary of outcome of interest. Must be 0 or 1.}

\item{s1}{The savings or benefit when am event is reclassified to a higher group by the new model}

\item{s2}{The benefit when a non-event is reclassified to a lower group}
}
\value{
A matrix of metrics for use within CI.classNRI
}
\description{
The function statistics.classNRI calculates the NRI metrics for reclassification of data already in classes. For use by CI.classNRI.
}

