% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rap.R
\name{statistics.raplot}
\alias{statistics.raplot}
\title{Reclassification metrics with calculated risk as inputs.}
\usage{
statistics.raplot(x1, x2, y, s1, s2, t)
}
\arguments{
\item{x1}{Calculated probabilities (eg through a logistic regression model) of the reference (null) model.  Must be between 0 & 1}

\item{x2}{Calculated probabilities (eg through a logistic regression model) of the new (alternative) model.   Must be between 0 & 1}

\item{y}{Binary of outcome of interest. Must be 0 or 1.}

\item{s1}{The savings or benefit when am event is reclassified to a higher group by the new model}

\item{s2}{The benefit when a non-event is reclassified to a lower even}

\item{t}{The risk threshold(s) for groups. eg t<-c(0,0.1,1) is a two group model with a threshold of 0.1 & t<-c(0,0.1,0.3,1)  is a three group model with thresholds at 0.1 and 0.3.}
}
\value{
A matrix of metrics for use within CI.raplot
}
\description{
The function statistics.raplot calculates the reclassification metrics. Used by CI.raplot.
}
\section{Further reading }{
  A useful reference for the weighted NRI is: Van Calster, B., Vickers, A. J., Pencina, M. J., Baker, S. G., Timmerman, D., & Steyerberg, E. W. (2013). Evaluation of markers and risk prediction models: overview of relationships between NRI and decision-analytic measures. Medical Decision Making : an International Journal of the Society for Medical Decision Making, 33(4), 490–501. doi:10.1177/0272989X12470757
}

